requirements.txt
aiogram==3.13.1
asyncpg-lite~=0.3.1.3
sqlalchemy[asyncio]>=2.0
redis[hiredis]>=5.0
aiohttp>=3.9
python-decouple>=3.8
yookassa>=3.0
apscheduler>=3.10
uvicorn>=0.30